# openapi.yaml
openapi: 3.0.3
info:
  title: Enterprise Multi-Agent Chat API
  version: 1.0.0
  description: REST and SSE contract for the enterprise multi-agent chat bot.
servers:
  - url: http://localhost:8080
paths:
  /api/chat/sse:
    post:
      summary: Send a user message and receive the agent answer as an SSE stream.
      operationId: chatStream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatMessageRequest'
      responses:
        '200':
          description: Streaming agent response chunks.
          content:
            text/event-stream:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/sessions:
    get:
      summary: List chat sessions for the current user.
      operationId: listSessions
      responses:
        '200':
          description: Session summaries.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SessionSummaryResponse'
    post:
      summary: Create a new chat session.
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '201':
          description: Created session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionSummaryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/sessions/{id}:
    delete:
      summary: Delete or clear a chat session.
      operationId: deleteSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '204':
          description: Session deleted.
        '404':
          $ref: '#/components/responses/NotFound'
  /api/sessions/{id}/messages:
    get:
      summary: List messages in a chat session.
      operationId: listMessages
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Session message history.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatMessageResponse'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  parameters:
    SessionId:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    AgentType:
      type: string
      enum:
        - SIMPLE_CHAT
        - RAG
        - SERVICE_SQL
    ChatMessageRequest:
      type: object
      required:
        - sessionId
        - userId
        - message
      properties:
        sessionId:
          type: string
        userId:
          type: string
        message:
          type: string
          minLength: 1
    CreateSessionRequest:
      type: object
      required:
        - userId
      properties:
        userId:
          type: string
        title:
          type: string
    SessionSummaryResponse:
      type: object
      required:
        - id
        - userId
        - title
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ChatMessageResponse:
      type: object
      required:
        - id
        - sessionId
        - role
        - content
        - createdAt
      properties:
        id:
          type: string
        sessionId:
          type: string
        role:
          type: string
          enum:
            - user
            - assistant
            - system
        content:
          type: string
        agentType:
          $ref: '#/components/schemas/AgentType'
        createdAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string